requirements.txt
fastapi
uvicorn
python-dotenv
langchain-google-genai
langchain==0.1.16
langchain-community==0.0.32
faiss-cpu
pypdf